+++ src/pages/index.tsx
@@ -1,24 +1,17 @@
-import Link from 'next/link';
-import type { NextPage } from 'next';
-import React from 'react';
-import styled from 'styled-components';
+import type { NextPage } from "next";
+import Link from "next/link";
+import styled from "styled-components";
+import Header from "../components/Header";
 
 const HomePage: NextPage = () => {
   return (
     <>
-      <Header>
-        <Link href='/'>
-          <Title>HAUS</Title>
-        </Link>
-        <Link href='/login'>
-          <p>login</p>
-        </Link>
-      </Header>
+      <Header />
       <Container>
-        <Link href='/pagination?page=1'>
+        <Link href="/pagination?page=1">
           <StyledLink>pagination</StyledLink>
         </Link>
-        <Link href='/infinite-scroll'>
+        <Link href="/infinite-scroll">
           <StyledLink>infinite scroll</StyledLink>
         </Link>
       </Container>
@@ -27,17 +20,6 @@ const HomePage: NextPage = () => {
 };
 
 export default HomePage;
-
-const Header = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 20px;
-`;
-
-const Title = styled.a`
-  font-size: 48px;
-`;
 
 const Container = styled.div`
   display: flex;
