+++ src/pages/infinite-scroll.tsx
@@ -1,22 +1,14 @@
-import Link from 'next/link';
-import type { NextPage } from 'next';
-import React from 'react';
-import styled from 'styled-components';
+import type { NextPage } from "next";
+import styled from "styled-components";
 
-import products from '../api/data/products.json';
-import ProductList from '../components/ProductList';
+import products from "../api/data/products.json";
+import Header from "../components/Header";
+import ProductList from "../components/ProductList";
 
 const InfiniteScrollPage: NextPage = () => {
   return (
     <>
-      <Header>
-        <Link href='/'>
-          <Title>HAUS</Title>
-        </Link>
-        <Link href='/login'>
-          <p>login</p>
-        </Link>
-      </Header>
+      <Header />
       <Container>
         <ProductList products={products} />
       </Container>
@@ -25,17 +17,6 @@ const InfiniteScrollPage: NextPage = () => {
 };
 
 export default InfiniteScrollPage;
-
-const Header = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 20px;
-`;
-
-const Title = styled.a`
-  font-size: 48px;
-`;
 
 const Container = styled.div`
   display: flex;
