+++ src/pages/pagination.tsx
@@ -1,12 +1,11 @@
-import { useRouter } from 'next/router';
-import Link from 'next/link';
-import type { NextPage } from 'next';
-import React from 'react';
-import styled from 'styled-components';
+import type { NextPage } from "next";
+import { useRouter } from "next/router";
+import styled from "styled-components";
 
-import products from '../api/data/products.json';
-import ProductList from '../components/ProductList';
-import Pagination from '../components/Pagination';
+import products from "../api/data/products.json";
+import Header from "../components/Header";
+import Pagination from "../components/Pagination";
+import ProductList from "../components/ProductList";
 
 const PaginationPage: NextPage = () => {
   const router = useRouter();
@@ -14,14 +13,7 @@ const PaginationPage: NextPage = () => {
 
   return (
     <>
-      <Header>
-        <Link href='/'>
-          <Title>HAUS</Title>
-        </Link>
-        <Link href='/login'>
-          <p>login</p>
-        </Link>
-      </Header>
+      <Header />
       <Container>
         <ProductList products={products.slice(0, 10)} />
         <Pagination />
@@ -31,17 +23,6 @@ const PaginationPage: NextPage = () => {
 };
 
 export default PaginationPage;
-
-const Header = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 20px;
-`;
-
-const Title = styled.a`
-  font-size: 48px;
-`;
 
 const Container = styled.div`
   display: flex;
