--- src/pages/products/[id].tsx
@@ -1,61 +1,0 @@
-import Link from 'next/link';
-import type { NextPage } from 'next';
-import React from 'react';
-import styled from 'styled-components';
-
-import products from '../../api/data/products.json';
-
-const ProductDetailPage: NextPage = () => {
-  const product = products[0];
-
-  return (
-    <>
-      <Header>
-        <Link href='/'>
-          <Title>HAUS</Title>
-        </Link>
-        <Link href='/login'>
-          <p>login</p>
-        </Link>
-      </Header>
-      <Thumbnail src={product.thumbnail ? product.thumbnail : '/defaultThumbnail.jpg'} />
-      <ProductInfoWrapper>
-        <Name>{product.name}</Name>
-        <Price>{product.price}원</Price>
-      </ProductInfoWrapper>
-    </>
-  );
-};
-
-export default ProductDetailPage;
-
-const Header = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 20px;
-`;
-
-const Title = styled.a`
-  font-size: 48px;
-`;
-
-const Thumbnail = styled.img`
-  width: 100%;
-  height: 420px;
-`;
-
-const ProductInfoWrapper = styled.div`
-  margin-top: 20px;
-  padding: 0 20px;
-`;
-
-const Name = styled.div`
-  font-size: 20px;
-  font-weight: bold;
-`;
-
-const Price = styled.div`
-  font-size: 18px;
-  margin-top: 8px;
-`;
